refactor: move util header to include
