+++ subprojects/libanim/include/libanim/util.h
