+++ subprojects/libanim/src/util/decrypt.c
@@ -1,4 +1,4 @@
-#include "util.h"
+#include "libanim/util.h"
 
 #include <string.h>
